pyproject.toml
[project]
name = "whisper-2-0"
version = "0.1.0"
description = "Audio transcription tool for Engineers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Audio processing
    "pydub>=0.25.1", # Audio manipulation
    # AI Services
    "openai>=1.0.0", # OpenAI API (includes Whisper API access)
    "groq>=0.4.0", # Groq API for fast Whisper transcription
    "google-generativeai>=0.7.1", # Gemini API
    # Configuration
    "python-dotenv>=1.0.0", # Loading .env files
    # Notion integration
    "notion-client>=2.0.0", # Notion API client
    # Notifications
    "plyer>=2.1.0", # Cross-platform notifications
    # Utilities
    "requests>=2.31.0", # HTTP requests for file downloads
    "ruff",
    "black",
]

[project.scripts]
whisper2 = "whisper_2_0.main:main"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.setuptools.packages.find]
where = ["src"]
